对用户、部门的操作按钮添加权限控制，对审核流程页面添加超级管理权限控制

See merge request !20

+++ src/main/java/com/meiren/web/acl/GroupModule.java
@@ -119,6 +119,11 @@ public class GroupModule extends BaseController {
         ApiResult result = new ApiResult();
         Map<String, Object> delMap = new HashMap<>();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasGroupAll(user)){
+                result.setError("您没有权限操作部门！");
+                return result;
+            }
             Long id = this.checkId(request);
             delMap.put("id", id);
             result = aclGroupService.deleteAclGroup(delMap);
@@ -190,6 +195,7 @@ public class GroupModule extends BaseController {
      * @param response
      * @return
      */
+    @AuthorityToken(needToken = {"meiren.acl.group.all"})
     @RequestMapping(value = "goTo/{type}", method = RequestMethod.GET)
     @ResponseBody
     public ModelAndView goTo(HttpServletRequest request, HttpServletResponse response,@PathVariable String type) {
@@ -227,6 +233,11 @@ public class GroupModule extends BaseController {
                              HttpServletResponse response, @PathVariable String type) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasGroupAll(user)){
+                result.setError("您没有权限操作部门！");
+                return result;
+            }
             Long initId = RequestUtil.getLong(request, "dataId");
             Long userId = RequestUtil.getLong(request, "selectedId");
             Long uid = RequestUtil.getLong(request, "uid");
@@ -310,6 +321,11 @@ public class GroupModule extends BaseController {
                                HttpServletResponse response, @PathVariable String type) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasGroupAll(user)){
+                result.setError("您没有权限操作部门！");
+                return result;
+            }
             Long initId = RequestUtil.getLong(request, "dataId");
             Long userId = RequestUtil.getLong(request, "selectedId");
             Long uid = RequestUtil.getLong(request, "uid");
@@ -392,6 +408,11 @@ public class GroupModule extends BaseController {
                              HttpServletResponse response, @PathVariable String type) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasGroupAll(user)){
+                result.setError("您没有权限操作部门！");
+                return result;
+            }
             Long initId = RequestUtil.getLong(request, "dataId");
             Long roleId = RequestUtil.getLong(request, "selectedId");
             Long uid = RequestUtil.getLong(request, "uid");
